Move general authentication to Experimental module
Removes the UndecidableInstances extension in the module containing the
HasServer instances.
